+++ src/main/java/com/raoulvdberge/refinedstorage/gui/grid/sorting/GridSortingInventoryTweaks.java
@@ -7,23 +7,28 @@ import invtweaks.api.InvTweaksAPI;
 import net.minecraftforge.fml.common.Loader;
 
 public class GridSortingInventoryTweaks extends GridSorting {
+    public static final String MOD_ID = "inventorytweaks";
+
     private InvTweaksAPI api = null;
 
     public GridSortingInventoryTweaks() {
         try {
             api = (InvTweaksAPI) Class.forName("invtweaks.forge.InvTweaksMod", true, Loader.instance().getModClassLoader()).getField("instance").get(null);
-        } catch (Exception ex) { }
+        } catch (Exception e) {
+            // NO OP
+        }
     }
-    
+
     @Override
-    public int compare(IGridStack o1, IGridStack o2) {
-        if (api != null && o1 instanceof GridStackItem && o2 instanceof GridStackItem) {
+    public int compare(IGridStack left, IGridStack right) {
+        if (api != null && left instanceof GridStackItem && right instanceof GridStackItem) {
             if (sortingDirection == NetworkNodeGrid.SORTING_DIRECTION_DESCENDING) {
-                return api.compareItems(((GridStackItem) o1).getStack(), ((GridStackItem) o2).getStack());
+                return api.compareItems(((GridStackItem) left).getStack(), ((GridStackItem) right).getStack());
             } else if (sortingDirection == NetworkNodeGrid.SORTING_DIRECTION_ASCENDING) {
-                return api.compareItems(((GridStackItem) o2).getStack(), ((GridStackItem) o1).getStack());
+                return api.compareItems(((GridStackItem) right).getStack(), ((GridStackItem) left).getStack());
             }
         }
+
         return 0;
     }
 }