+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Refined Storage Changelog
 
 ### 1.5.15
+- Updated Forge to 2444 (MC 1.12.1) (raoulvdberge)
+- Added InventoryTweaks Grid sorting (cooliojazz)
 - Added CTM integration for Disk Manipulator (raoulvdberge)
 
 ### 1.5.14